+++ internal/edgeos/data_test.go
@@ -24,8 +24,7 @@ func logIt(w io.Writer, s string) {
 
 func shuffleArray(slice []string) {
 	rand.Seed(time.Now().UnixNano())
-	n := len(slice)
-	for i := n - 1; i > 0; i-- {
+	for i := len(slice) - 1; i > 0; i-- {
 		j := rand.Intn(i + 1)
 		slice[i], slice[j] = slice[j], slice[i]
 	}